Fix async execution and write multi-step searches
Resolved an issue with the asynchronous execution: it is now actually
asynchronous and not a direct call to another function. There appears to
be a bug in the unit testing library that doesn't support usage of
`process.nextTick`, but for now this has been worked around by using the
less efficient `setTimeout callback, 0` function.

The search algorithm also now supports multiple step paths and caveats
identified around that have been documented in the code (regarding state
copying). Unit tests for the above have been written.