Fix frontier and explored lists searching
Fixed the code that checks whether the current state has been seen
before, either in the frontier or the explored lists. A sort of
quick-and-dirty fix using the `underscore` library's `any` function.
This will probably be the hottest of the framework and so may require
some attention in the future.